Добавил SearchHistory в код

+++ commons/src/main/java/org/parsingbot/commons/entity/SearchHistory.java
@@ -19,7 +19,7 @@ public class SearchHistory {
     @GeneratedValue(strategy = GenerationType.SEQUENCE)
     private Integer id;
     private String vacancyName;
-    private Integer numberOfVacancies;
+    private Long numberOfVacancies;
     private String keywords;
 
     @ManyToOne(fetch = FetchType.LAZY, optional = false)
